requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
python-dotenv==1.0.0
openai>=1.40.0
orjson==3.9.10
numpy==1.24.3
streamlit==1.28.1
datasets==2.14.5
plotly==5.17.0
pandas==2.0.3
requests==2.31.0
